refactor(navigation): implement worker navigation graph
Add a new worker navigation graph to manage worker-related screens
and their navigation flows. This refactor enhances the modularity
of the navigation structure, allowing for better organization and
maintenance of worker-specific routes and components.
